requirements.txt
torch
torchvision
pandas
nltk
scikit-learn
Pillow
tensorboard
tqdm 
Flask 